+++ FinanceControl.WebApi/Controllers/AreaController.cs
@@ -1,10 +1,15 @@
 ﻿using FinanceControl.Domain.Interfaces.Service;
+using FinanceControl.Services.Extensions;
 using FinanceControl.Services.Services;
 using FinanceControl.Services.Validations;
 using FinanceControl.Shared.Dtos.Request;
+using FinanceControl.WebApi.Controllers.Base;
+using FinanceControl.WebApi.Extensions;
+using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.ComponentModel.DataAnnotations;
 
 namespace FinanceControl.WebApi.Controllers
 {
@@ -14,15 +19,23 @@ namespace FinanceControl.WebApi.Controllers
     public class AreaController : BaseController
     {
         private readonly IAreaService _areaService;
+        private readonly IValidator<CreateAreaRequestDto> _createAreaValidator;
+        private readonly IValidator<UpdateAreaRequestDto> _updateAreaValidator;
 
-        public AreaController(IAreaService areaService)
+        public AreaController(IAreaService areaService, IValidator<CreateAreaRequestDto> createAreaValidator, IValidator<UpdateAreaRequestDto> updateAreaValidator)
         {
             _areaService = areaService;
+            _createAreaValidator = createAreaValidator;
+            _updateAreaValidator = updateAreaValidator;
         }
 
         [HttpPost]
         public async Task<IActionResult> CreateAreaAsync([FromBody]CreateAreaRequestDto requestDto)
         {
+            var validationResult = _createAreaValidator.Validate(requestDto);
+            if (validationResult.ToActionResult() is { } errorResult)
+                return errorResult;
+
             var userId = GetUserId();
 
             var result = await _areaService.CreateAreaAsync(requestDto, userId);
@@ -34,8 +47,12 @@ namespace FinanceControl.WebApi.Controllers
         }
 
         [HttpGet("all/{budgetId:int}")]
-        public async Task<IActionResult> GetAllAreaAsync([FromQuery] int budgetId)
+        public async Task<IActionResult> GetAllAreaAsync([FromRoute] int budgetId)
         {
+            var validationResult = this.ValidatePositiveId(budgetId, "budgetId");
+            if (validationResult is not null)
+                return validationResult;
+
             var userId = GetUserId();
 
             var result = await _areaService.GetAllAreasAsync(budgetId, userId);
@@ -45,15 +62,27 @@ namespace FinanceControl.WebApi.Controllers
         [HttpGet("by-id/{id:int}")]
         public async Task<IActionResult> GetAreaByIdAsync([FromRoute] int id)
         {
+            var validationResult = this.ValidatePositiveId(id, "id");
+            if (validationResult is not null)
+                return validationResult;
+
             var userId = GetUserId();
 
             var result = await _areaService.GetAreaByIdAync(id, userId);
+
+            if (result == null)
+                return NotFound("Area not found.");
+
             return Ok(result);
         }
 
         [HttpPatch]
         public async Task<IActionResult> UpdateAreaAsync([FromBody] UpdateAreaRequestDto requestDto)
         {
+            var validationResult = _updateAreaValidator.Validate(requestDto);
+            if (validationResult.ToActionResult() is { } errorResult)
+                return errorResult;
+
             var userId = GetUserId();
 
             var result = await _areaService.UpdateAreaAsync(requestDto, userId);
@@ -67,6 +96,10 @@ namespace FinanceControl.WebApi.Controllers
         [HttpDelete("{id:int}")]
         public async Task<IActionResult> DeleteAreaAsync([FromRoute] int id)
         {
+            var validationResult = this.ValidatePositiveId(id, "id");
+            if (validationResult is not null)
+                return validationResult;
+
             var userId = GetUserId();
             var result = await _areaService.DeleteAreaAsync(id, userId);
