+++ FinanceControl.WebApi/Controllers/CategoryController.cs
@@ -2,6 +2,8 @@
 using FinanceControl.Services.Extensions;
 using FinanceControl.Shared.Dtos.Request;
 using FinanceControl.Shared.Models;
+using FinanceControl.WebApi.Controllers.Base;
+using FinanceControl.WebApi.Extensions;
 using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -28,8 +30,8 @@ namespace FinanceControl.WebApi.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateCategoryAsync([FromBody]CreateCategoryRequestDto requestDto)
         {
-            var validatorResult = _createCategoryValidator.Validate(requestDto);
-            if (validatorResult.ToActionResult() is { } errorResult)
+            var validatonResult = _createCategoryValidator.Validate(requestDto);
+            if (validatonResult.ToActionResult() is { } errorResult)
                 return errorResult;
 
             var userId = GetUserId();
@@ -53,8 +55,8 @@ namespace FinanceControl.WebApi.Controllers
         [HttpPatch("by-id")]
         public async Task<IActionResult> UpdateCategoryByIdAsync([FromBody] UpdateCategoryRequestDto requestDto)
         {
-            var validatorResult = _updateCategoryValidator.Validate(requestDto);
-            if (validatorResult.ToActionResult() is { } errorResult)
+            var validatonResult = _updateCategoryValidator.Validate(requestDto);
+            if (validatonResult.ToActionResult() is { } errorResult)
                 return errorResult;
 
             var userId = GetUserId();
@@ -70,6 +72,10 @@ namespace FinanceControl.WebApi.Controllers
         [HttpDelete("{id:int}")]
         public async Task<IActionResult> DeleteCategoryByIdAsync([FromRoute]int id)
         {
+            var validationResult = this.ValidatePositiveId(id, "id");
+            if (validationResult is not null)
+                return validationResult;
+
             var userId = GetUserId();
             var result = await _categoryService.DeleteCategoryByIdAsync(id, userId);
